Implement dumb git http protocol
OpenChariot now documents and supports the git "dumb" http protocol for
anonymous clones. This is useful, because users behind firewalls which
block git:// (common for large corperations) will be able to perform
clones.

This feature is, for now, turned on by default, and can be disabled in
ocutil-getconfig as documented.